==> scripts/10_printpage_mystery-1.Rmd <==
---
output: tufte::tufte_handout
---

```{r setup, include=FALSE}
library(tufte)
# invalidate cache when the tufte version changes
knitr::opts_chunk$set(tidy = FALSE, cache.extra = packageVersion('tufte'))
options(htmltools.dir.version = FALSE)
```

\thispagestyle{empty}
\vspace*{-4\baselineskip}
\section{Mystery Question 1}

\vspace{5in}


\section{Mystery Question 1}




\clearpage
\thispagestyle{empty}
\vspace*{-4\baselineskip}
\section{Mystery Question 2}

\vspace{5in}


\section{Mystery Question 2}





\clearpage
\thispagestyle{empty}
\vspace*{-4\baselineskip}
\section{Mystery Question 3}

\vspace{5in}


\section{Mystery Question 3}





\clearpage
\thispagestyle{empty}
\vspace*{-4\baselineskip}
\section{Mystery Question 4}

\vspace{5in}


\section{Mystery Question 4}
